+++ mw/index.ts
@@ -7,7 +7,7 @@ import {instances} from './util';
 declare const $STYLE: string;
 
 // 每次新增插件都需要修改这里
-const baseVersion = '3.5',
+const baseVersion = '3.9',
 	addons = ['lint'];
 
 mw.loader.addStyleTag($STYLE);